PLUG-103 fix: BLOG003-카테고리 변수타입 수정

+++ src/main/java/com/justdo/plug/blog/domain/category/Category.java
@@ -15,7 +15,7 @@ import lombok.Setter;
 public class Category extends BaseTimeEntity{
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int category_id;
+    private Long category_id;
 
     @Column(nullable = false)
     private String name;
